backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
tensorflow==2.18.0
opencv-python==4.9.0.80
pillow==10.2.0
numpy==1.26.3
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2
